requirements.txt
bottle==0.12.17
dnspython==1.16.0
nltk==3.4.5
numpy==1.17.4
pymongo==3.9.0
python-dotenv==0.10.3
six==1.13.0
